move files around to more logical places
